Modified UI for UpdatePreferences.
