+++ src/components/UpdatePreferences.js
@@ -8,10 +8,10 @@ import app from "../firebase"
 import firebase from "firebase/app"
 import 'firebase/firestore'
 import { auth } from "../firebase"
+import { useAuth } from "../contexts/AuthContext"
 
 const useStyles = makeStyles({
   root: {
-    width: 300,
   },
 });
 
@@ -20,11 +20,21 @@ function valuetext(value) {
 }
 
 export default function UpdatePreferences() {
+  const [userDataDoc, setUserDataDoc] = useState("")
   const [error, setError] = useState("")
   const classes = useStyles();
   const [value, setValue] = React.useState([0, 100]);
+  const {fetchUserDocument } = useAuth();
   const [currValue = fetchAgeRange(), setCurrValue] = React.useState();
   
+  async function getUserData() {
+	  const doc = await fetchUserDocument();
+	  return doc;
+  }
+  if (!userDataDoc) {
+	  getUserData().then(doc => setUserDataDoc(doc.data()));
+  }
+
   const handleChange = (event, newValue) => {
     setValue(newValue);
   };
@@ -55,25 +65,99 @@ export default function UpdatePreferences() {
   
   return (
     <>
+    <div class="container">
+   <div class="row">
+      <div class="col-md-12">
+         <div id="content" class="content content-full-width">
+            {/* <!-- begin profile --> */}
+            <div class="profile">
+               <div class="profile-header">
+                <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"></link>
+                     
+                  {/* <!-- BEGIN profile-header-cover --> */}
+                  <div class="profile-header-cover">
+                      <div class="profile-header-cover-edit pull-right">
+                        <a href="#" class="btn btn-sm btn-info mb-3"> 
+                          <span class="glyphicon glyphicon-edit" ></span>
+                        </a>
+                      </div>
+                  </div>
+                  {/* <!-- END profile-header-cover -->
+                  <!-- BEGIN profile-header-content --> */}
+                  <div class="profile-header-content">
+                     {/* <!-- BEGIN profile-header-img --> */}
+                     <div class="profile-header-img">
+                        <img src="https://bootdey.com/img/Content/avatar/avatar3.png" alt=""></img>
+                        
+                     </div>
+                     {/* <!-- END profile-header-img -->
+                     <!-- BEGIN profile-header-info --> */}
+                     <div class="profile-header-info">
+                     { userDataDoc &&
+                        <h4 class="m-t-10 m-b-5">{userDataDoc["Name"]}</h4>
+                     }
+                        <div>
+                          {userDataDoc && userDataDoc["Profession"].join(',')}
+                       </div>
+                        <a href="/profile-interests" class="btn btn-sm btn-dark mb-2">Go back to Profile</a>
+                     </div>
+                     <div class="profile-header-img-edit">
+                     <a href="#" class="btn btn-info btn-sm"> 
+                     <span class="glyphicon glyphicon-edit"></span> Edit
+                      </a>
+                     </div>
+                     
+                     {/* <a href="/profile-picture-edit" class="btn btn-sm btn-info mb-2">Edit Profile Picture</a> */}
+                     {/* <!-- END profile-header-info --> */}
+                  </div>
+                  {/* <!-- END profile-header-content -->
+                  <!-- BEGIN profile-header-tab --> */}
+                  <ul class="profile-header-tab nav nav-tabs">
+                  <li class="nav-item"> 
+                        <Link to="/update-profile" class="nav-link " data-toggle="tab">
+                            ACCOUNT
+                        </Link>
+                    </li>
+                    <li class="nav-item"> 
+                        <Link to="/update-profile-about" class="nav-link" data-toggle="tab">
+                            ABOUT
+                        </Link>
+                    </li>
+                     <li class="nav-item">
+                        <Link to="/update-profile-interests" class="nav-link" data-toggle="tab">INTERESTS</Link></li>
+                    <li class="nav-item"><Link to="/update-preferences" class="nav-link  active show" data-toggle="tab">PREFERENCES</Link>
+                     </li> 
+                     <li class="nav-item"><Link to="/update-profile-events" class="nav-link" data-toggle="tab">EVENTS</Link>
+                     </li>
+                     
+                  </ul>
+                  {/* <!-- END profile-header-tab --> */}
+                  </div>
+            </div>
+            {/* <!-- end profile -->
+            <!-- begin profile-content -->*/}
       <Card>
         <Card.Body>
           <h2 className="text-center mb-4">Update Age Range</h2>
 		  {currValue ? 
-		  <h2 className="text-center mb-4">Current age range set to : {currValue[0]} - {currValue[1]}</h2>
+		  <h4 className="mb-4">Current age range set to : {currValue[0]} - {currValue[1]}</h4>
 		  : null}
-          {error && <Alert variant="danger">{error}</Alert>}
-            <div className={classes.root}>
-			<Typography id="range-slider" gutterBottom>
-				Select age range
-			</Typography>
-			<Slider
-				value={value}
-				onChange={handleChange}
-				valueLabelDisplay="auto"
-				aria-labelledby="range-slider"
-				getAriaValueText={valuetext}
-			/>
-			</div>
+      {error && <Alert variant="danger">{error}</Alert>}
+        <div className={classes.root}>
+          <Typography id="range-slider" gutterBottom>
+            <h4> Select age range</h4>
+          </Typography>
+          <Slider
+            value={value}
+            onChange={handleChange}
+            valueLabelDisplay="auto"
+            aria-labelledby="range-slider"
+            getAriaValueText={valuetext}
+            width="100%"
+            step={1}
+            marks
+          />
+        </div>
             <Button className="w-100" onClick={() => {handleUpdate(value);}}>
               Update
             </Button>
@@ -81,6 +165,10 @@ export default function UpdatePreferences() {
       </Card>
       <div className="w-100 text-center mt-2">
         <Link to="/">Cancel</Link>
+      </div>
+      </div>
+      </div>
+      </div>
       </div>
     </>
   )
